Removed transactions (#31)
## What Changed
Removed transactions from all stores and firestore library components
Refactored "coins" to stickerInstances
Cleaned up some unused code 
Enabled videos to play in room even if in test mode without
streamPlaybackID


## Context
We previously had a transaction and energy system. Each sticker push to
database was also accompanied with a transaction, that would be
processed by the server, and contribute to streamers energy. In this old
world, stickers were sometimes called coins which was confusing. This
was removed in the server, and now we are removing it in the client and
refactoring to represent the new world

Additionally, when testing around I noticed that the videos werent
playing in the rooms. Realized that the videoPlayer expected a
streamPlaybackID which is not required if it's in test mode, so
refactored to accommodate this change.

## Validation
Verified that videos play within the rooms, even without playbackID, and
stickers still work.


https://github.com/user-attachments/assets/774765a3-6c01-44d1-b59c-6e95b1b43985


## Relevant issues
Closes #5

---------

Co-authored-by: bhavik singh <bhavik@bhaviksworkbook.lan>