Enabling presence (#32)
# What Changed
Added server/.env.local, which will be preferred when running server 
Updated presence code in server to work
Sent presence and showed num online in client 

# Context
We had a half working presence system previously. I re-activated it, by
updating the client to actually send the presence heartbeat, and the
server to correctly process heartbeats and update numOnline in room.

In doing so, I realized that it's frustrating for the server to be
working with the PROD firebase database. So I added a .env.local file to
the server, and had the server read from it if available. This should be
safe in prod which uses Fly secrets instead of .env files.

# Validation

Tried it out with the local server, where everything seems to work. Once
merged, will deploy fly.


https://github.com/user-attachments/assets/58b73f9d-408a-467f-90d4-653ef180f37e


# Relevant issues
Closes #6

---------

Co-authored-by: bhavik singh <bhavik@bhaviksworkbook.lan>