Mobile functionality  (#45)
## Changes

Added functionality to make the website mobile friendly (no video, just
chat) and some other tweaks.
- On Mobile sizes, we use defaultRoom for all rooms, displaying only
chat.
- On Mobile sizes, special layout is used to stack footer and content
- Tweaked domRing to render nicely on mobile.
- Added back button to account page 
- Make it so that the offline rooms can be clicked into, to set stage
for issue #7

## Context

Context is that previously the site was just trying to render the entire
desktop page on mobile, which was very broken. Now, we have a special
mobile view.

## Testing

Not tested in app, will do after deploy. Tested in inspector.


https://github.com/user-attachments/assets/3fd0f569-7330-434f-9cbc-3a094173aae6

## Relevant Issues

- Closes #14 
- Closes #11

---------

Co-authored-by: bhavik singh <bhavik@bhaviksworkbook.lan>
Co-authored-by: bhavik singh <bhavik@bhaviks-WorkBook.local>