Ability to make rooms dynamically (#56)
## Changes
- Added a room creator in settings. 
- Added a room name / color updater in settings 
- Removed all the mux information from settings

## Context

For the stream on Dec 29, we are enabling isto for up to 50 streamers.
For this, we will be hosting a zoom session where streamas can come in
and create rooms. To enable this, we need to let each streamer create
their own room. Additionally, mux is no longer relevant since we're
using stream - so we now can remove all that information from settings.

## Testing

Validated on dev branch that I can create rooms.

## Relevant Issues

- Partially move towards #42

---------

Co-authored-by: bhavik singh <bhavik@bhaviks-WorkBook.local>
Co-authored-by: bhavik singh <bhavik@mac.lan>
Co-authored-by: bhavik singh <bhavik@bhaviksworkbook.lan>