Post test fixes + autoscanner (#57)
## Changes

Multiple quick fixes and changes.
- Slowed down presence, so that it is much more calm when there are lots
of people online
- Fixed bug in streamgate, where it was not disconnecting the client on
unmount. Used AI for this, so worth double checking
- Refactored a bunch of button styling
- Added a new "autoscanner" functionality to jump between rooms
- UI tweaks for the way rooms render, to support both dark room colors
as well as support long room names

## Context

After the 30 person test stream, we found a bunch of issues that we
fixed
- Error where the client is complaining that too many sockets are open,
implying web sockets were not closed.
- Streamers added rooms with long names and dark backgorunds 

Additionally, for the install at nguyen wahed we need auto-scanning
functionality so the rooms can be navigating unmaintained.

## Testing

Tested that it works with multiple rooms

https://github.com/user-attachments/assets/bd5fd775-9dd0-43ed-b59e-c1a732abe169

Tested that it correctly loops on home infinitely when no rooms are
active

https://github.com/user-attachments/assets/3424ce73-8bfe-4bcb-a928-42ac4607e561




## Relevant Issues

Note, there are still some race conditions with the websockets not FULLY
being cleaned up when the room is being changed. Likely race condition
where room is changed before the getClient callback is succesful,
resulting in errors like: `too many requests for user: bhaviksingh,
check response headers for more information.","isWSFailure":false` which
is different from before. but still open.

---------

Co-authored-by: bhavik singh <bhavik@bhaviks-WorkBook.local>
Co-authored-by: bhavik singh <bhavik@mac.lan>
Co-authored-by: bhavik singh <bhavik@bhaviksworkbook.lan>